Mesh-Derived Reeb Graphs

These are some of the Reeb graphs used by Tu &al (2019; Fig. 8) to benchmark the multi-pass and single-pass algorithms. The original meshes were obtained from the AIM@SHAPE Shape Repository, while the authors computed Reeb graphs using a custom C++ implementation.

Usage

david
buddha
topology
flower

Format

Objects of class reeb_graph.

Source

http://visionair.ge.imati.cnr.it/ontologies/shapes/ (AIM@SHAPE; defunct) https://github.com/USFDataVisualization/ReebGraphPairing